Constrain window more effectively.
Also don't let the window become smaller than x, because `x>window`
implies that the next sample within the window may not be a fibonacci
number, and then we might miss it (and consequently all the values that
followed what we should have found right after).

And rename `step` to `phi`.